Don't report success when the SNS publish failed; cast bigint arrays correctly
Two real findings from review.

**publishMessage swallowed its failures.** It catches and returns, and returns
early when AF_TOPIC_ARN is unset, so a failed publish still recorded status
'created'. The Lambda was never asked to build the quiz, and the round then sat on
"Generating links…" indefinitely with nothing in the UI or the row saying why.

It now returns whether it published (still never throws — a publish failure must
not fail a create that already succeeded). Setup checks the result and marks the
row 'failed' with a retryable message instead of 'created'. Existing callers that
ignore the return value are unaffected, so /api/quiz-sessions keeps its current
behaviour; it has the same blind spot and can adopt this when someone touches it.

Not addressed (and unchanged by this commit): a Lambda that fails *after*
receiving the message still can't report back, so a row can read 'created' while
the quiz build failed — only CloudWatch knows. That needs a callback or a
reconciler, and is noted in the context file.

**::int[] against bigint columns.** cycles/route.ts and teacher-feedback-batches.ts
cast to ::int[] while session.id and batch.id are bigints — a mismatch that throws
"integer out of range" past 2^31. Max ids are ~18k so there was no live risk, but
the cast should agree with the column. Pinned by a test.

+++ src/app/api/teacher-feedback/cycles/route.ts
@@ -87,7 +87,7 @@ async function resolveSessionLinks(
     portal_link: string | null;
     meta_data: { admin_testing_link?: string } | null;
   }>(
-    `SELECT id, platform_id, portal_link, meta_data FROM session WHERE id = ANY($1::int[])`,
+    `SELECT id, platform_id, portal_link, meta_data FROM session WHERE id = ANY($1::bigint[])`,
     [sessionPks]
   );
   for (const s of sessionRows) {
